+++ .eyas.config.js
@@ -46,9 +46,6 @@ module.exports = {
 
 	// defaults to current platform
 	outputs: {
-		// windows: true,
-		// mac: true,
-		// linux: true,
 		// expires: 720
 	}
 };